MAESTRO Phase 6: continued compilation fixes — 49/67 (73%)
Remaining 18 failures are deep domain-specific issues:
- AIRace (3): missing RaceData context, AI struct method mismatches
- AudioEngine (2): OpenAL function stubs needed
- Collision (5): KartObjectProxy includes, KartCollisionInfo, HitboxGroup
- Physics (3): pointer casts, constructor syntax, method calls
- RaceEngine (3): course enum names, BattleManager, Time type
- Rendering (4): pointer arithmetic, CourseObjActor API, singleton patterns

+++ data/src/game/Rendering/CourseObj.cpp
@@ -195,7 +195,7 @@ void CourseObjActor_UpdateWithRotation(CourseObjActor* self, s32 variantIdx) {
         s32 sign = (variantIdx < lodMin) ? -1 : ((lodMin < variantIdx) ? 1 : 0);
         s32 absSign = (sign < 0) ? -sign : sign;
         u32 newOffset = 0x348 + (absSign * 0x0C);
-        self->courseDataPtr = reinterpret_cast<u32>(reinterpret_cast<u8*>(sRenderSettings) + newOffset);
+        self->courseDataPtr = reinterpret_cast<uintptr_t>(reinterpret_cast<u8*>(sRenderSettings) + newOffset);
 
         inLodRange = false;
